move broker's ingress handler and metrics to pkg
